requirements.txt
Django >=2.2.0, <2.3.0
djangorestframework >=3.9.2, <3.10.0
coverage >=4.5.3, <4.6.0
psycopg2-binary >= 2.8.6, <2.9
flake8 >= 3.8.4, < 3.9
django-cors-headers >= 3.5.0, <3.6.0
